Remove customProfiles boolean
Problem: The customProfiles boolean was used to decide whether or not we should call delete on custom profiles. However, if customProfiles dict was empty, the boolean was set to false and we never finished cleaning up profiles. We don't need this boolean, we will always look to see if any profiles should be deleted.

Solution: Remove this boolean check to allow an empty customProfiles struct to still be processed and all remaining profiles get cleaned up.