Modified parameter passed in to call delete_unused_ssl_profiles
Problem:  Need access to the F5CloudServiceManager object in
the delete_unused_ssl_profiles function defined in CCCL. This
allows the function to make use of the virtual objects previously
created by apply_ltm_config().  Otherwise, the virtuals would
have to be re-retrieved from the Big-IP.

Solution: Passed in the manager object to the call which has the
required information already stored.  This information is needed
to help prune the delete list for profiles (to remove those held
by whitelisted virtuals).